requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.7
django-bootstrap4==2.3.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.2.0
PyJWT==1.7.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
django-debug-toolbar
djangorestframework-guardian